### 2021-09-17 #464 @christophsax
Submission tsbox: Class-Agnostic Time Series
Submitting Author Name:  Christoph Sax
Submitting Author Github Handle: @christophsax
Repository:  https://github.com/christophsax/tsbox
Version submitted: 3.1.9001
Submission type: Stats
Badge grade: silver
Reviewers: @chamberlinc, @brunaw, @nunesmatt
Version accepted: TBD

---

```
Package: tsbox
Title: Class-Agnostic Time Series
Version: 0.3.1.9001
Authors@R: person("Christoph", "Sax", email = "christoph.sax@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-7192-7044"))
Description: Time series toolkit with identical behavior for all
  time series classes: 'ts','xts', 'data.frame', 'data.table', 'tibble', 'zoo',
  'timeSeries', 'tsibble', 'tis' or 'irts'. Also converts reliably between these classes.
    data.table (>= 1.10.0),
    anytime
    forecast,
    seasonal,
    dygraphs,
    xts,
    scales,
    tsibble (>= 0.8.2),
    tsibbledata,
    tibbletime,
    tseries,
    units,
    zoo,
    tis,
    timeSeries,
    nycflights13,
    imputeTS,
    spelling
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
Config/testthat/parallel: true
```

- [X] A pre-submission inquiry has been approved in [issue#457](#457)

Anyone who works with time series. Many statistical packages require time series to be in a certain object (ts, xts, tsibble, data.frame). tsbox facilitates the conversion between these objects. It also provides a general toolkit that works the same way with all time series classes. {tsbox} is also mentioned in the rOpenSci Statistical Software Peer Review Section on Time Series.

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

In the *rOpenSci* classification, this package is *An improvement on other implementations of similar algorithms in **R***. Many time series packages, e.g., [zoo](https://CRAN.R-project.org/package=zoo) or [tsibble](https://CRAN.R-project.org/package=tsibble) contain converter functions from one class to another. They often convert from their class to `ts` objects and back, but lack converters to other time series class.

In most cases, tsbox transforms an object into an augmented `data.table`. And uses the `data.table` infrastructure for efficient joining and reshaping. After computation, it restores the original input class. This restoring feature is
was also used in the `xts::reclass()` function of the [xts](https://CRAN.R-project.org/package=xts) package.

-    Please include hyperlinked references to all other relevant software.

[data.table](https://CRAN.R-project.org/package=data.table): For efficient joining and reshaping
[xts](https://CRAN.R-project.org/package=xts): Similar reclassing mechanism
[tsibble](https://CRAN.R-project.org/package=tsibble): Tidy Temporal Data Frames and Tools

I probably need some advice on this. I think that tsbox complies with most standards that are applicable. Generality of usage is a particular feature that should qualify the package for silver.

The most outstanding point is probably the one on generality:

> Have a demonstrated *generality* of usage beyond one single envisioned use case.

The package facilitates work with time series in general. It can also be used to ease the burden of object testing and time series conversion for other time series packages.

- [X] I/we have read the [guide for authors](https://devguide.ropensci.org/guide-for-authors.html) and [rOpenSci packaging guide](https://devguide.ropensci.org/building.html).

    *Already published on CRAN*.
