+++ man/tsbox-package.Rd
@@ -26,8 +26,8 @@ use the same tsbox-command for any time series classes.
 The best way to start is to check out the package
 \href{https://www.tsbox.help}{website}.
 
-In the \emph{ropensci} classification, this package is *An improvement on other
-implementations of similar algorithms in \strong{R}. Many time series packages,
+In the \emph{ropensci} classification, this package is \emph{An improvement on other
+implementations of similar algorithms in \strong{R}}. Many time series packages,
 e.g., \href{https://CRAN.R-project.org/package=zoo}{zoo} or
 \href{https://CRAN.R-project.org/package=tsibble}{tsibble} contain converter
 functions from one class to another. They often convert from their class
@@ -36,7 +36,7 @@ to \code{ts} objects and back, but lack converters to other time series class.
 In most cases, tsbox transforms an object into an augmented \code{data.table}. And
 uses the \code{data.table} infrastructure for efficient joining and reshaping. After
 computation, it restores the original input class. This restoring feature is
-was also used in the `xts::reclass() function of the
+was also used in the \code{xts::reclass()} function of the
 \href{https://CRAN.R-project.org/package=xts}{xts} package.
 }
 \author{
