### 2021-09-18 8827dcd
test for ts_plot()
