### 2021-09-18 #464 @christophsax
#### R CMD check

There seems to be an issue with the package seasonal on your system. The package is special since it uses x13binary which downloads pre-built binaries from CRAN. Usually this works fine and on most CRAN platforms.

Is there anything special about your system? There was an issue for Mac M1.

If you want to check, this should works:

```r
library(seasonal)
seas(AirPassengers)
checkX13()
```

But I don't want you to drag into this. I can also remove the few tests that are using seasonal, which should then resolve the issue?

#### Coverage

Perhaps this is just because of the failing R CMD check? If not, is there a way to see what ist the measured coverage? I had 81.2% on my system, which passed the threshold (see below).

#### Report on my side

If I run `pkgcheck::pkgcheck(".")` on my side, I get:
```r
#### Coverage

✔ Package uses 'roxygen2'.
✔ Package has a 'contributing.md' file.
✔ Package has a 'CITATION' file.
✔ Package has a 'codemeta.json' file.
✔ All functions have examples.
✔ Package has at least one HTML vignette
✔ Package 'DESCRIPTION' has a URL field.
✔ Package 'DESCRIPTION' has a BugReports field.
✔ Package is already on CRAN.
✔ Package has continuous integration checks.
✔ Package coverage is 81.2%.
✔ R CMD check found no errors.
✔ R CMD check found no warnings.
✔ All applicable standards [v0.0.1] have been documented in this package.

ℹ Current status:
✔ This package may be submitted.
```
