### 2021-10-19 #464 @rkillick
The goodpractice() checks came back with:
- Use <- instead of =
ts_frequency.R L120
- Lines longer than 80 characters:
ts_bind.R L58, L62
ts_first_of_period.R L12, L13, L18
test-ts_frequency.R L33

I had problems running covr() but see that it is passing on the checks above.  Please fix the above and I'll look to assign reviewers as these are small changes.
