+++ tests/testthat/test-tsibble.R
@@ -33,8 +33,9 @@ test_that("tsibble back-conversion works properly", {
   weather <- nycflights13::weather %>%
     select(origin, time_hour, temp, humid, precip)
   weather_tsbl <- as_tsibble(weather, key = origin, index = time_hour)
-  weather_tsbl %>%
+  ans <- weather_tsbl %>%
     ts_default()
+  expect_s3_class(ans, "tbl_ts")
 })
 
 
