+++ R/ts_first_of_period.R
@@ -15,13 +15,6 @@
 #' )
 #' ts_first_of_period(x)
 #' ts_first_of_period(ts_lag(ts_df(austres), "14 days"))
-#' \donttest{
-#' x <- ts_lag(data.frame(
-#'   time = seq(as.POSIXct("1970-01-01"), length.out = 10, by = "10 sec"),
-#'   value = rnorm(10)
-#' ), "3 sec")
-#' ts_first_of_period(x)
-#' }
 #' @export
 ts_first_of_period <- function(x) {
   ts_apply(x, dts_first_of_period)
