### 2021-12-06 #464 @chamberlinc
@rkillick @christophsax , thank you for the opportunity to review, and my apologies that this has taken so long. Below is my review with some comments. I also have uploaded the knitted html notebook that details the tests I ran at https://github.com/chamberlinc/tsbox-review. The code I used to test the package functions is also available in this repo.

Errors encountered in `ts_plot()`, `ts_xts()` and `ts_long()`. I am including the structure of the data frame I was working with in the code below:
Errors with `ts_xts()` and `ts_long()`:

```
> Conn_discharge_DO <- dataRetrieval::readNWISuv(
+   siteNumbers = "01193050",
+   parameterCd = c("00060", "00300"),
+   startDate = "2019-01-01",
+   endDate = "2021-01-01"
+ )
>
> str(Conn_discharge_DO)
'data.frame':	70519 obs. of  8 variables:
 $ agency_cd       : chr  "USGS" "USGS" "USGS" "USGS" ...
 $ site_no         : chr  "01193050" "01193050" "01193050" "01193050" ...
 $ dateTime        : POSIXct, format: "2019-01-01 05:00:00" "2019-01-01 05:15:00" "2019-01-01 05:30:00" "2019-01-01 05:45:00" ...
 $ X_00060_00000   : num  39200 41300 40400 40400 41100 43000 41800 42300 41700 42400 ...
 $ X_00060_00000_cd: chr  "A" "A" "A" "A" ...
 $ X_00300_00000   : num  13.9 13.9 13.9 14 14 14 14 14 14 14 ...
 $ X_00300_00000_cd: chr  "A" "A" "A" "A" ...
 $ tz_cd           : chr  "UTC" "UTC" "UTC" "UTC" ...
 - attr(*, "url")= chr "https://nwis.waterservices.usgs.gov/nwis/iv/?site=01193050&format=waterml,1.1&ParameterCd=00060,00300&startDT=2"| __truncated__
 - attr(*, "siteInfo")='data.frame':	1 obs. of  13 variables:
  ..$ station_nm          : chr "CONNECTICUT RIVER AT MIDDLE HADDAM, CT"
  ..$ site_no             : chr "01193050"
  ..$ agency_cd           : chr "USGS"
  ..$ timeZoneOffset      : chr "-05:00"
  ..$ timeZoneAbbreviation: chr "EST"
  ..$ dec_lat_va          : num 41.5
  ..$ dec_lon_va          : num -72.6
  ..$ srs                 : chr "EPSG:4326"
  ..$ siteTypeCd          : chr "ST-TS"
  ..$ hucCd               : chr "01080205"
  ..$ stateCd             : chr "09"
  ..$ countyCd            : chr "09007"
  ..$ network             : chr "NWIS"
 - attr(*, "variableInfo")='data.frame':	2 obs. of  7 variables:
  ..$ variableCode       : chr [1:2] "00060" "00300"
  ..$ variableName       : chr [1:2] "Streamflow, ft&#179;/s" "Dissolved oxygen, water, unfiltered, mg/L"
  ..$ variableDescription: chr [1:2] "Discharge, cubic feet per second" "Dissolved oxygen, water, unfiltered, milligrams per liter"
  ..$ valueType          : chr [1:2] "Derived Value" "Derived Value"
  ..$ unit               : chr [1:2] "ft3/s" "mg/l"
  ..$ options            : chr [1:2] "" ""
  ..$ noDataValue        : logi [1:2] NA NA
 - attr(*, "disclaimer")= chr "Provisional data are subject to revision. Go to http://waterdata.usgs.gov/nwis/help/?provisional for more information."
 - attr(*, "statisticInfo")='data.frame':	1 obs. of  2 variables:
  ..$ statisticCd  : chr "00000"
  ..$ statisticName: chr ""
 - attr(*, "queryTime")= POSIXct[1:1], format: "2021-12-06 11:53:57"
>
> # Try viewing both parameters together
> try(str(ts_xts(Conn_discharge_DO)))
more than one [value] column detected after [time] - using the outermost.
Are you using a wide data frame? To convert, use 'ts_long()'.

[time]: 'dateTime' [value]: 'X_00300_00000'
Error : cannot allocate vector of size 4.5 Gb
> try(str(ts_xts(ts_long(Conn_discharge_DO))))
[id] columns left of [time] column: 'agency_cd', 'site_no'
[time]: 'dateTime'
Error : 'value' column [value] is not numeric.
```

Errors with `ts_plot()`:
```
> character_date_dat <- data.frame(
+   DateTime = c(
+   "2011-11-11 11:11:11", "2012-12-12 12:12:12", "2021-09-25 20:07:00"
+   ),
+   Value = c(1, 2, 8)
+ )
> try({ts_plot(character_date_dat)})
[time]: 'DateTime' [value]: 'Value'
Error in colnamesInt(x, neworder, check_dups = FALSE) :
  argument specifying columns specify non existing column(s): cols[3]='Value'
> # I am downloading hydrological data timeseries from the USGS. This is my most frequent way of accessing timeseries data. Data comes as a data frame.
> Eno_discharge <- dataRetrieval::readNWISuv(
+   siteNumbers = "02085070",
+   parameterCd = "00060",
+   startDate = "2019-01-01",
+   endDate = "2021-01-01"
+ )
> # View the structure of the data frame
> str(Eno_discharge)
'data.frame':	70267 obs. of  6 variables:
 $ agency_cd       : chr  "USGS" "USGS" "USGS" "USGS" ...
 $ site_no         : chr  "02085070" "02085070" "02085070" "02085070" ...
 $ dateTime        : POSIXct, format: "2019-01-01 05:00:00" "2019-01-01 05:15:00" "2019-01-01 05:30:00" "2019-01-01 05:45:00" ...
 $ X_00060_00000   : num  424 421 424 428 428 428 428 428 435 435 ...
 $ X_00060_00000_cd: chr  "A" "A" "A" "A" ...
 $ tz_cd           : chr  "UTC" "UTC" "UTC" "UTC" ...
 - attr(*, "url")= chr "https://nwis.waterservices.usgs.gov/nwis/iv/?site=02085070&format=waterml,1.1&ParameterCd=00060&startDT=2019-01"| __truncated__
 - attr(*, "siteInfo")='data.frame':	1 obs. of  13 variables:
  ..$ station_nm          : chr "ENO RIVER NEAR DURHAM, NC"
  ..$ site_no             : chr "02085070"
  ..$ agency_cd           : chr "USGS"
  ..$ timeZoneOffset      : chr "-05:00"
  ..$ timeZoneAbbreviation: chr "EST"
  ..$ dec_lat_va          : num 36.1
  ..$ dec_lon_va          : num -78.9
  ..$ srs                 : chr "EPSG:4326"
  ..$ siteTypeCd          : chr "ST"
  ..$ hucCd               : chr "03020201"
  ..$ stateCd             : chr "37"
  ..$ countyCd            : chr "37063"
  ..$ network             : chr "NWIS"
 - attr(*, "variableInfo")='data.frame':	1 obs. of  7 variables:
  ..$ variableCode       : chr "00060"
  ..$ variableName       : chr "Streamflow, ft&#179;/s"
  ..$ variableDescription: chr "Discharge, cubic feet per second"
  ..$ valueType          : chr "Derived Value"
  ..$ unit               : chr "ft3/s"
  ..$ options            : chr ""
  ..$ noDataValue        : logi NA
 - attr(*, "disclaimer")= chr "Provisional data are subject to revision. Go to http://waterdata.usgs.gov/nwis/help/?provisional for more information."
 - attr(*, "statisticInfo")='data.frame':	1 obs. of  2 variables:
  ..$ statisticCd  : chr "00000"
  ..$ statisticName: chr ""
 - attr(*, "queryTime")= POSIXct[1:1], format: "2021-12-06 11:59:00"
> # Try viewing the discharge data using the ts_plot function
> try({ts_plot(Eno_discharge)})
[time]: 'dateTime' [value]: 'X_00060_00000'
Error in setnames(x, c(cid, ctime, cvalue), c("id", "time", "value")) :
  Items of 'old' not found in column names: [X_00060_00000]. Consider skip_absent=TRUE.
> try({ts_plot(Eno_discharge, skip_absent=TRUE)})
Error in FUN(X[[i]], ...) : ts_boxable(x) is not TRUE
```

```
> library(testthat)
> library(tsbox)
>
> test_check("tsbox")
Starting 2 test processes
== Skipped tests ===============================================================
* empty test (1)
== Warnings ====================================================================
-- Warning (test-ts_first_of_period.R:18:3): ts_first_of_period works ----------
no non-missing arguments to min; returning Inf
 1. tsbox::ts_first_of_period(x)
 2. tsbox::ts_apply(x, dts_first_of_period)
 3. tsbox:::ts_apply_dts(ts_dts(x), fun, ...)
 4. tsbox:::fun(x, ...)
 8. base::NextMethod("[")
== Failed tests ================================================================
-- Error (test-ts_first_of_period.R:18:3): ts_first_of_period works ------------
Error in `max(which(time <= smry$start)):min(which(time >= smry$end))`: result would be too long a vector
    x
 1. \-tsbox::ts_first_of_period(x)
 2.   \-tsbox::ts_apply(x, dts_first_of_period)
 3.     \-tsbox:::ts_apply_dts(ts_dts(x), fun, ...)
 4.       \-tsbox:::fun(x, ...)
 5.         +-time[(max(which(time <= smry$start)):min(which(time >= smry$end)))]
 6.         +-base::`[.POSIXct`(...)
 7.         | \-base::.POSIXct(NextMethod("["), attr(x, "tzone"), oldClass(x))
 8.         \-base::NextMethod("[")
1 error x | 1 warning x | 0 notes √
```

---

First, I enjoyed the opportunity to review this code. This is the first code review I have done for rOpenSci, and I hope the comments below are helpful.

I like the general purpose of this package and agree there is probably a lot of use for a package that allows easy conversion between R object types. The documentation I think is all ok, though I think it could be improved with a little more description of the expected structure of data. I did not have any issues with installation. There was one test that failed on my machine (see details in the index.nb.html file in the linked repo, <https://github.com/chamberlinc/tsbox-review>). The code snippets in the three vignettes all worked on my machine.

I will preface my comments by saying I am not an experienced user of data.tables, so following the code was a little tricky for me. I usually work with timeseries as data frames, and I was having trouble following the code and figuring out which columns were being assigned as data, timestamps, or identifiers. This caused a few issues when I had data frames of multiple concurrent timeseries and was getting error messages indicating that the ordering of the columns somehow may have mattered. When working with data frames that only had one timeseries, the functions all worked simply and, it seemed, as intended.

I was mostly experimenting with the `ts_plot()`, `ts_long()`, and `ts_xts()` functions. `ts_xts()` produced an error while working with the data frame that had multiple timeseries. Following the prompting, I tried using `ts_long()`, however this also produced an error that I did not understand, because it seemed to be accessing columns that I did not intend. I didn't see any way to input the correct column names to the function. Once I manually pivoted the data frame using `pivot_longer()` though, I was able to use `ts_xts()` and it performed everything correctly I believe. `ts_plot()` produced error messages that it might have mattered how the columns were organized, and that would have been helpful to know. I was not able to get this function to work for several formats of data frame data, even when following the suggested prompts in the error messages.

The code I used to access data and experiment is provided as well at the linked repo. I hope this is helpful!
