### 2022-04-18 305b660
extend vignette on expected data frame structure
