+++ tests/testthat/test-dirty.R
@@ -4,12 +4,12 @@ library(tsbox)
 
 test_that("works with df with improper col classes", {
   library(dplyr)
-  x.chr <- ts_tbl(mdeaths) %>%
+  x.chr <- ts_tbl(mdeaths) |>
     mutate(time = as.character(time))
 
   expect_s3_class(ts_ts(x.chr), "ts")
 
-  x.fct <- ts_tbl(mdeaths) %>%
+  x.fct <- ts_tbl(mdeaths) |>
     mutate(time = as.factor(as.character(time)))
 
   expect_s3_class(ts_ts(x.fct), "ts")
