+++ vignettes/tsbox.Rmd
@@ -314,19 +314,19 @@ ts_long(dta_wide)
 
 ## Using tsbox in a dplyr / pipe workflow
 
-tsbox works well with tibbles and with `%>%`, so it can be nicely integrated
+tsbox works well with tibbles and with the pipe (`|>` or `%>%`), so it can be nicely integrated
 into a dplyr workflow:
 
 ```r
 library(dplyr)
 library(nycflights13)
-dta <- weather %>%
-  select(origin, time = time_hour, temp, humid, precip) %>%
+dta <- weather |>
+  select(origin, time = time_hour, temp, humid, precip) |>
   ts_long()
 
-dta %>%
-  filter(id == "temp") %>%
-  ts_trend() %>%
+dta |>
+  filter(id == "temp") |>
+  ts_trend() |>
   ts_plot()
 ```
 
