+++ DESCRIPTION
@@ -2,7 +2,10 @@ Package: tsbox
 Type: Package
 Title: Class-Agnostic Time Series
 Version: 0.3.1.9001
-Authors@R: person("Christoph", "Sax", email = "christoph.sax@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-7192-7044"))
+Authors@R: c(
+    person("Christoph", "Sax", email = "christoph.sax@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-7192-7044")),
+    person("Cathy", "Chamberlin", role = c("rev"))
+    )
 Description: Time series toolkit with identical behavior for all
   time series classes: 'ts','xts', 'data.frame', 'data.table', 'tibble', 'zoo',
   'timeSeries', 'tsibble', 'tis' or 'irts'. Also converts reliably between these classes.
