### 2022-04-19 ce524e6
Fix rolling join behind ts_first_of_period
