### 2022-04-19 fa48a70
fix NA handling in rolling join in data.table
