+++ NAMESPACE
@@ -89,6 +89,7 @@ importFrom(data.table,"setkey")
 importFrom(data.table,"setkeyv")
 importFrom(data.table,"setnames")
 importFrom(data.table,"setorder")
+importFrom(data.table,"shift")
 importFrom(data.table,"tstrsplit")
 importFrom(grDevices,bmp)
 importFrom(grDevices,dev.off)
