### 2022-04-19 0e64a33
ts_first_of_period works with POSIXct fix #210
