### 2022-04-19 #464 @christophsax
First of all, please apologize for my late answer. I originally planned to wait for the second review and lost track afterward.

@chamberlinc, thanks again for your very helpful review! I dealt with your points in [PR #211](https://github.com/christophsax/tsbox/pull/211).

Most of your problems arose when applying the functions to non-standard data frames, especially wide ones. In some cases, the messages already pointed to the use use of `ts_long()`, but the error messages were not too helpful. In [PR #211](https://github.com/christophsax/tsbox/pull/211), I tried to address some of these problems. The error messages are improved, and ts_long() can handle some more complex cases. It works now with a toy example of your data set.

A bug in ts_plot() also caused a weird error message. This bug is now resolved as well.

To sum up, I addressed the following three points from your review:

### description of the expected data structure

> I think it could be improved with a little more description of the expected structure of data.

Extended paragraph on data frames: <https://www.tsbox.help/articles/tsbox.html#time-series-in-data-frames-1>

### Failing test

> There was one test that failed on my machine (see details in the index.nb.html file in the linked repo, <https://github.com/chamberlinc/tsbox-review>).

The error and the warning did not occur on my system, but the computation in `ts_first_of_period()` did not work as expected. I added a new test and a fix. Both are green now on my systems (local and GHA tests in the repository).

### Error Messages on wide data frames

> This caused a few issues when I had data frames of multiple concurrent timeseries and was getting error messages indicating that the ordering of the columns somehow may have mattered.

I improved some of the error messages when functions were applied to wide data structures.

``` r
library(nycflights13)
suppressPackageStartupMessages(library(dplyr))
library(tsbox)
packageVersion("tsbox")
#> [1] '0.3.1.9002'

d3 <- weather |>
  select(origin, time_hour, temp, humid, precip)
d3
#> # A tibble: 26,115 × 5
#>    origin time_hour            temp humid precip
#>    <chr>  <dttm>              <dbl> <dbl>  <dbl>
#>  1 EWR    2013-01-01 01:00:00  39.0  59.4      0
#>  2 EWR    2013-01-01 02:00:00  39.0  61.6      0
#>  3 EWR    2013-01-01 03:00:00  39.0  64.4      0
#>  4 EWR    2013-01-01 04:00:00  39.9  62.2      0
#>  5 EWR    2013-01-01 05:00:00  39.0  64.4      0
#>  6 EWR    2013-01-01 06:00:00  37.9  67.2      0
#>  7 EWR    2013-01-01 07:00:00  39.0  64.4      0
#>  8 EWR    2013-01-01 08:00:00  39.9  62.2      0
#>  9 EWR    2013-01-01 09:00:00  39.9  62.2      0
#> 10 EWR    2013-01-01 10:00:00  41    59.6      0
#> # … with 26,105 more rows
```

This is a wide data frame, so tsbox errors with a meaningful message:

``` r
ts_ts(d3)
#> Found numeric [id] column(s): 'temp', 'humid'.
#> Are you using a wide data frame? To convert, use 'ts_long()'.
#> Convert column(s) to character or factor to silence this message.
#> [time]: 'time_hour' [value]: 'precip'
#> Error: series has no regular pattern
```

This also works if the order is different

``` r
d4 <- weather |>
  select(origin, temp, humid, precip, time_hour)
d4
#> # A tibble: 26,115 × 5
#>    origin  temp humid precip time_hour
#>    <chr>  <dbl> <dbl>  <dbl> <dttm>
#>  1 EWR     39.0  59.4      0 2013-01-01 01:00:00
#>  2 EWR     39.0  61.6      0 2013-01-01 02:00:00
#>  3 EWR     39.0  64.4      0 2013-01-01 03:00:00
#>  4 EWR     39.9  62.2      0 2013-01-01 04:00:00
#>  5 EWR     39.0  64.4      0 2013-01-01 05:00:00
#>  6 EWR     37.9  67.2      0 2013-01-01 06:00:00
#>  7 EWR     39.0  64.4      0 2013-01-01 07:00:00
#>  8 EWR     39.9  62.2      0 2013-01-01 08:00:00
#>  9 EWR     39.9  62.2      0 2013-01-01 09:00:00
#> 10 EWR     41    59.6      0 2013-01-01 10:00:00
#> # … with 26,105 more rows

ts_ts(d4)
#> Found numeric [id] column(s): 'temp', 'humid'.
#> Are you using a wide data frame? To convert, use 'ts_long()'.
#> Convert column(s) to character or factor to silence this message.
#> [time]: 'time_hour' [value]: 'precip'
#> Error: series has no regular pattern
```

Using `ts_plot()` on these malformed data frames has returned the following nonsensical error:

    [time]: 'time_hour' [value]: 'precip'
    Error in setnames(x, c(cid, ctime, cvalue), c("id", "time", "value")) :
      Items of 'old' not found in column names: [precip]. Consider skip_absent=TRUE.
    >

The `skip_absent=TRUE` is a data.table suggestion, not meant for use in tsbox. I fixed this in commit 50e694b8. The message now makes more sense:

``` r
ts_plot(d3)
#> Found numeric [id] column(s): 'temp', 'humid'.
#> Are you using a wide data frame? To convert, use 'ts_long()'.
#> Convert column(s) to character or factor to silence this message.
#>
#> [time]: 'time_hour' [value]: 'precip'
#> too many series. Only showing the first 20.
```

reflecting the fact that we have a large number of absurdly short time series (since value columns were classified as id columns)

I also tried to improve a bit on your examples. I construct a data frame that looks similar to yours:

``` r
# > str(Conn_discharge_DO)
# 'data.frame':   70519 obs. of  8 variables:
#  $ agency_cd       : chr  "USGS" "USGS" "USGS" "USGS" ...
#  $ site_no         : chr  "01193050" "01193050" "01193050" "01193050" ...
#  $ dateTime        : POSIXct, format: "2019-01-01 05:00:00" "2019-01-01 05:15:00" "2019-01-01 05:30:00" "2019-01-01 05:45:00" ...
#  $ X_00060_00000   : num  39200 41300 40400 40400 41100 43000 41800 42300 41700 42400 ...
#  $ X_00060_00000_cd: chr  "A" "A" "A" "A" ...

Conn_discharge_DO <- tibble(
  agency_cd = c("USGS", "USGS", "USGS", "USGS"),
  site_no = c("01193050", "01193050", "01193050", "01193050"),
  dateTime = as.POSIXct(c("2019-01-01 05:00:00", "2019-01-01 05:15:00", "2019-01-01 05:30:00", "2019-01-01 05:45:00")),
  unit = "ft3/s",
  X_00060_00000 = 1:4,
  noDataValue = NA,
  X_00060_000002 = 11:14,
  X_00060_00000_cd = "A"
)
```

tsbox now detects `X_00060_00000` as an id column. (As it says in the documentation, it starts on the right). And it suggests using `ts_long()`.

``` r
ts_xts(Conn_discharge_DO)
#> Found numeric [id] column(s): 'X_00060_00000'.
#> Are you using a wide data frame? To convert, use 'ts_long()'.
#> Convert column(s) to character or factor to silence this message.
#> [time]: 'dateTime' [value]: 'X_00060_000002'
#> Loading required namespace: xts
#>                     USGS_01193050_ft3/s_1_NA_A USGS_01193050_ft3/s_2_NA_A
#> 2019-01-01 05:00:00                         11                         NA
#> 2019-01-01 05:15:00                         NA                         12
#> 2019-01-01 05:30:00                         NA                         NA
#> 2019-01-01 05:45:00                         NA                         NA
#>                     USGS_01193050_ft3/s_3_NA_A USGS_01193050_ft3/s_4_NA_A
#> 2019-01-01 05:00:00                         NA                         NA
#> 2019-01-01 05:15:00                         NA                         NA
#> 2019-01-01 05:30:00                         13                         NA
#> 2019-01-01 05:45:00                         NA                         14
```

`ts_long()` works now on this. It classifies character and factor columns right of the time column as id columns (with a message).

``` r
ts_long(Conn_discharge_DO)
#> found columns right to the [time] column that will be treated as [id] columns (character or factor): 'unit', 'X_00060_00000_cd'.
#> Additional [id] column(s): 'agency_cd', 'site_no', 'unit', 'X_00060_00000_cd'
#> [time]: 'dateTime'
#> # A tibble: 12 × 7
#>    agency_cd site_no  unit  X_00060_00000_cd id        dateTime            value
#>    <chr>     <chr>    <chr> <chr>            <chr>     <dttm>              <int>
#>  1 USGS      01193050 ft3/s A                X_00060_… 2019-01-01 05:00:00     1
#>  2 USGS      01193050 ft3/s A                X_00060_… 2019-01-01 05:15:00     2
#>  3 USGS      01193050 ft3/s A                X_00060_… 2019-01-01 05:30:00     3
#>  4 USGS      01193050 ft3/s A                X_00060_… 2019-01-01 05:45:00     4
#>  5 USGS      01193050 ft3/s A                noDataVa… 2019-01-01 05:00:00    NA
#>  6 USGS      01193050 ft3/s A                noDataVa… 2019-01-01 05:15:00    NA
#>  7 USGS      01193050 ft3/s A                noDataVa… 2019-01-01 05:30:00    NA
#>  8 USGS      01193050 ft3/s A                noDataVa… 2019-01-01 05:45:00    NA
#>  9 USGS      01193050 ft3/s A                X_00060_… 2019-01-01 05:00:00    11
#> 10 USGS      01193050 ft3/s A                X_00060_… 2019-01-01 05:15:00    12
#> 11 USGS      01193050 ft3/s A                X_00060_… 2019-01-01 05:30:00    13
#> 12 USGS      01193050 ft3/s A                X_00060_… 2019-01-01 05:45:00    14
```
