### 2022-05-28 #464 @chamberlinc
Hello Christoph,
Apologies for the delay again. These updates I think are very good. The functions worked as expected this time with the data I typically work with. I had a few automated tests fail this time though, one because I have an older version of R that does not support the |> operator, and the rest all seemingly because of some issue accessing the function ts_dts.

The details of the automated test output you can find here: https://github.com/chamberlinc/tsbox-review/blob/master/tsbox_R2.md. The rest of my review is below:

## Package Review

-   [X ] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   [ X] **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README

-   [X ] **Installation instructions:** for the development version of package and any non-standard dependencies in README

-   [X ] **Vignette(s):** demonstrating major functionality that runs successfully locally
> Caveat: I am running R 4.0.5 on my machine, and so I do not have native pipes. For the tsbox vignette, the code did work, but I first had to replace all instances |> with %>%. I’d recommend using the older syntax to be compatable with more versions of R.

-   [X ] **Function Documentation:** for all exported functions
> I did not check all of the functions’ documentation, but the subset I checked looked good.

-   [X ] **Examples:** (that run successfully locally) for all exported functions
> Again, only checked a subset, but all worked.

-   [X \ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   [X ] **Installation:** Installation succeeds as documented.
-   [X ] **Functionality:** Any functional claims of the software been confirmed.
-   [X ] **Performance:** Any performance claims of the software been confirmed.
-   [  ] **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine. Several checks and tests still seem to fail, at least one of which is likely due to native pipes with my version of R. The other has to do with the ts_dts function.

    sessionInfo()

    ## Running under: Windows 10 x64 (build 19044)
    ##
    ## Matrix products: default
    ##
    ## [5] LC_TIME=English_United States.1252
    ##
    ## [1] stats     graphics  grDevices utils     datasets  methods   base
    ##
    ## [13] xfun_0.31       digest_0.6.29   rlang_1.0.2     evaluate_0.15

    pkg_dir <- "C:/Users/Cathy/Documents/PeerReviewing/tsbox/R2/tsbox/"
    try(devtools::check(pkg_dir))

    ##                                            ^

    devtools::test(pkg_dir)

    ## Output removed for clarity. View at https://github.com/chamberlinc/tsbox-review/blob/master/tsbox_R2.md ##

-   [X ] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 3

-   [X ] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

The functions now work as expected for me for the types of data I typically use. Thanks for the updates! I found the messages on the ts_long function to be very helpful.
