### 2022-07-19 b1df63d
avoid date classes not getting respected by stopifnot, fixes #213
