### 2022-08-08 bb32ef8
use magrittr a little longer
