+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
 library(tsbox)
+library(dplyr)
 
 test_check("tsbox")
