### 2022-08-08 #464 @christophsax
Very sorry for the slow turnaround.

### Native Pipe

I changed the native pipe (`|>`) back to the magrittr one (`%>%`), so this should now work on your system with the latest version (`0.3.1.9003`).

I introduced the native pipe deliberately in the vignettes since I always wanted to keep the dependency graph very simple (tsbox only imports data.table and anytime) and wanted to make the vignettes available for non-magrittr/dplyr users. But I see that there is a trade-off between this aim and the support for older versions of R. The native pipe was introduced in R 4.1 on 2021-05-18 and is now available for almost 15 months. With more time passing, only a few older versions will be around. I changed it back for now but will go to the native pipe again at some point in the future.

### Other Test Errors

I have a hard time reproducing them. First, all the tests run through on my system (macOS, R release) and all systems on [GitHub Actions](https://github.com/christophsax/tsbox/actions/runs/2815299570) (ubuntu-20.04, release, devel), (windows-latest, release), (macOS-latest, release). So I wonder if this is another problem related to the old R version?

Your error says:

```r
 Error in `UseMethod("ts_dts")`: no applicable method for 'ts_dts' applied to an object of class "data.frame"
```

But there clearly is such a method:

https://github.com/christophsax/tsbox/blob/dd0447c66405da6101419c31b0fe3f0cffdfadc1/R/to_from_data.frame.R#L14-L18

If you still encounter the error in the tests, could you try to run one the failing example outside of the tests?
Does this fail? It really shouldn't.

```r
library(tsbox)
with_id <- wo_id <- ts_df(mdeaths)
with_id$id <- "mdeaths"
ts_c(wo_id, with_id)
```

Thank you so much for your work! I guess we are alomost done now. I will answer quickly this time :-)
