### 2022-08-22 #464 @chamberlinc
Hello, and sorry for my slow turn around this time.

I think that plan for the pipes makes the most sense. Unfortunately there are a significant number of users who don't have control over updating their own versions of R because of the IT protocols of their companies or agencies. But, given enough time I agree it will be less of a problem and decreasing dependencies is a good goal.

The error that was occurring previously has resolved itself on my computer. However, I am now getting a new error with the test `test-ts_first_of_period`:

```
Warning (test-ts_first_of_period.R:19:3): ts_first_of_period works
no non-missing arguments to min; returning Inf
Backtrace:
 1. tsbox::ts_first_of_period(x)
      at test-ts_first_of_period.R:19:2
 2. tsbox::ts_apply(x, dts_first_of_period)
      at tsbox/R/ts_first_of_period.R:20:2
 3. tsbox::ts_apply_dts(ts_dts(x), fun, ...)
      at tsbox/R/ts_apply.R:44:2
 4. tsbox fun(x, ...)
      at tsbox/R/ts_apply.R:18:4
 8. base::NextMethod("[")

Error (test-ts_first_of_period.R:19:3): ts_first_of_period works
Error in `max(which(time <= smry$start)):min(which(time >= smry$end))`: result would be too long a vector
Backtrace:
 1. tsbox::ts_first_of_period(x)
      at test-ts_first_of_period.R:19:2
 2. tsbox::ts_apply(x, dts_first_of_period)
      at tsbox/R/ts_first_of_period.R:20:2
 3. tsbox::ts_apply_dts(ts_dts(x), fun, ...)
      at tsbox/R/ts_apply.R:44:2
 4. tsbox fun(x, ...)
      at tsbox/R/ts_apply.R:18:4
 8. base::NextMethod("[")
 ```
I tried stepping through the test code outside of the test function, and I got the same error. Going into the `dts_first_of_period` function, the error seems to be because `which(time >= smry$end))` produces `integer(0)`.

Other than that though, everything worked well.
