### 2022-08-24 475b35e
prolong template series so it works in western time zones
