+++ NEWS.md
@@ -1,4 +1,4 @@
-# tsbox 0.3.1.9003
+# tsbox 0.3.1.9004
 
 ## Features
 
