### 2022-08-24 #464 @christophsax
@chamberlinc, thanks a lot!

The error in `ts_first_of_period()` was time zone dependent, that's why I haven't spotted it. I added another test and fixed the problem. It should work now with the latest version on GitHub (`0.3.1.9004`).
