### 2022-08-24 #464 @chamberlinc
@christophsax , all of the tests pass now, congrats! I was also able to recreate the problem with 'no applicable method' from the last round - when I run the tests while in a project they fail, but running the tests outside of a project they pass. I'm now assuming this was just a problem of working directories somehow.
