### 2022-09-30 #464 @nunesmatt
@rkillick @christophsax
Thank you for thinking of me to review this package (this is my first ROpenSci review). Please find some comments on the package below. The full script of this review can be found at my github page [here](https://github.com/nunesmatt/tsbox-review/tsbox-revew.md).

tsbox review
================
nunesmatt
30/09/2022

  - [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to
    review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

  - [x] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines. Estimated hours spent reviewing: 7

  - [x] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package
    DESCRIPTION file.

-----

The package aims to supply a structure/class-agnostic approach to time series analysis to circumvent some of the arguably tedious user conversion between classes before analysis tasks. In doing so, the functionality aims at generality – I feel this is a useful package in terms of functionality and scope for a wide range of end-users. The package is well-documented and the three vignettes all run smoothly.

Some comments on the package are below. Note that these are suggestions rather than fixes per se:

1.  **Package checking**. When checking the package, I get:

    dts_first_of_period: no visible binding for global variable ‘time.orig’
    ##   time.orig

Whilst this is not an error, one suggestion might be to define `time.orig` as `NULL` (like is done for `has.value`) at the start of the
function or before the subsetting in the `dts_first_of_period` function (contained within `ts_first_of_period.R`).

2.  **good practice**. A few things came up when using `gp`. Many code lines come up as 81 characters (as opposed to suggested lengths of ![\\leq 80](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;%5Cleq%2080
    "\\leq 80")), which might be due to line endings etc. `guess_dts.R` has a suggestion of using `vapply` rather than `sapply`. These are left to the developer to tackle if desired.

3.  **Error message about `ts_boxable(x)`**. I wonder whether a more informative error message could be shown when `ts_boxable(x)` is not `TRUE`. For example, whilst I realise that the package is designed for existing time series classes, using a `tsbox` function on a vector or inappropriate object results in the error, but the user may not know how to fix the issue, especially as `ts_boxable()` is primarily internal. Perhaps the message could point to the expected structure of the argument of the main package functions, e.g. “…please ensure the object `x` has components…” or “Is the object of class x, y or z?”

4.  **Comparison with `timetk`**. During playing with this package, I mainly focussed on the conversion functions (`ts_tbl()`,
    `ts_long()`,`ts_ts()`etc), as, whilst delving into testing datasets, I came across the `timetk` package, which at first glance aims to do a similar job to `tsbox`.

All functions in `tsbox` worked as expected, and some arguably have a more natural output than `timetk` (see full review document).

The package developer might want to consider an additional function handling the `zooreg` class for completeness.

5.  Whilst users are probably familiar with usage, the developer might want to consider adding examples to show functionality of `ts_prcomp` and `ts_forecast` under the `ts_examples` documentation file.

6.  I noticed that there is a spelling mistake of *explicit* in the error messages for the `guess_time` and `guess_value` functions.
