### 2022-12-08 #464 @mpadge
@christophsax Sorry that this review has taken such an extraordinarily long time. One of our editors has had to step back for a bit, so I'll take over editing the submission from here. [Your last comment](https://github.com/ropensci/software-review/issues/464#issuecomment-1263552573) indicated that you stil intended to respond to the reviews. An update would be appreciated, to help  us move this submission towards approval. Thanks :smile:
