### 2022-12-26 d3e26af
sapply -> vapply
