### 2022-12-26 1d38ef4
stopifnot(ts_boxable(x)) -> check_ts_boxable(x)
