+++ R/to_from_tis.R
@@ -28,7 +28,7 @@ ts_dts.tis <- function(x) {
 #' @name ts_ts
 #' @export
 ts_tis <- function(x) {
-  stopifnot(ts_boxable(x))
+  check_ts_boxable(x)
   if (relevant_class(x) == "tis") {
     return(x)
   }
