+++ R/to_from_zoo.R
@@ -26,7 +26,7 @@ ts_dts.zoo <- function(x) {
 #' @name ts_ts
 #' @export
 ts_zoo <- function(x) {
-  stopifnot(ts_boxable(x))
+  check_ts_boxable(x)
   if (relevant_class(x) == "zoo") {
     return(x)
   }
