+++ R/to_from_ts.R
@@ -149,7 +149,7 @@ ts_dts.ts <- function(x) {
 #' @importFrom stats resid time ts tsp as.formula var prcomp start tsp<- window
 #' @importFrom utils getFromNamespace browseURL relist
 ts_ts <- function(x) {
-  stopifnot(ts_boxable(x))
+  check_ts_boxable(x)
   if (relevant_class(x) == "ts") {
     return(x)
   }
