### 2022-12-26 83a96b9
Support for 'zooreg' time series
