### 2022-12-26 48bc938
export and document check_ts_boxable()
