+++ man/ts_boxable.Rd
@@ -2,9 +2,12 @@
 % Please edit documentation in R/to_from_all.R
 \name{ts_boxable}
 \alias{ts_boxable}
+\alias{check_ts_boxable}
 \title{Test if an Object is ts-Boxable}
 \usage{
 ts_boxable(x)
+
+check_ts_boxable(x)
 }
 \arguments{
 \item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo},
@@ -12,7 +15,8 @@ ts_boxable(x)
 \code{irts} or \code{timeSeries}.}
 }
 \value{
-logical, either \code{TRUE} or \code{FALSE}
+logical, either \code{TRUE} or \code{FALSE}. \code{check_ts_boxable()} fails if not
+\code{TRUE}
 }
 \description{
 Mainly used internally.
