+++ NAMESPACE
@@ -16,6 +16,7 @@ export("%ts*%")
 export("%ts+%")
 export("%ts-%")
 export("%ts/%")
+export(check_ts_boxable)
 export(colors_tsbox)
 export(copy_class)
 export(load_suggested)
