### 2022-12-26 #464 @christophsax
Thank you very much for your review, @nunesmatt! And thanks to you both for your patience, @nunesmatt and @mpadge!

I tried to answer your points as much as possible - except for the line length of the SRR comments and point 5., where I think that the point is already fulfilled.

> 1.  **Package checking**.

Fixed, thanks! (https://github.com/christophsax/tsbox/pull/216/commits/a9650f8e7ca2b86d4024ddece71db9e7ba7c185e)

> 2.  **good practice**. A few things came up when using `gp`. Many code lines come up as 81 characters (as opposed to suggested lengths of ![\\leq 80].

These are all SRR stats comments, e.g., `@srrstats {G2.4c}`. I had them broken once at 80 characters, but this caused another problem. So, I would leave them like this if you don't insist on it.

> `guess_dts.R` has a suggestion of using `vapply` rather than `sapply`.

Fixed, thanks!
(https://github.com/christophsax/tsbox/pull/216/commits/d3e26afa63403d81b58e679d660e5d5e6979224e)

> **Error message about `ts_boxable(x)`**.

Substituted exsiting `stopifnot(ts_boxable(x))` by the already existing but rarely used `check_ts_boxable(x)`. Result is a more verbose error message with hint to further information:

Previously:

    > ts_ts(1:2)
    Error in ts_ts(1:2) : ts_boxable(x) is not TRUE

Now:

```r
ts_ts(1:2)
    Error: object is of non-ts-boxable class(es) 'integer'. See `?ts_ts`.
```

> 4.  **Comparison with `timetk`**.

Added a paragraph on 'related packages' to the introductory vignette.
(https://github.com/christophsax/tsbox/pull/216/commits/5a774a3b5e8c224c61d04c68f6be3bafd3e1739c)

> The package developer might want to consider an additional function
handling the `zooreg` class for completeness.

Implemented. You can do now, e.g.,

```r
ts_zooreg(mdeaths)
```

> 5.  Whilst users are probably familiar with usage, the developer might
    want to consider adding examples to show functionality of
    `ts_prcomp` and `ts_forecast` under the `ts_examples` documentation
    file.

There are examples both for `ts_prcomp()` and `ts_forecast`, already in the [CRAN version](https://www.rdocumentation.org/packages/tsbox/versions/0.3.1/topics/ts_examples). Is it confusing that they are shown within `ts_plot()`?

> 6.  I noticed that there is a spelling mistake of *explicit* in the
    error messages for the `guess_time` and `guess_value` functions.

Done.
