### 2023-01-20 #464 @nunesmatt
Review 2 for `ts_box`:

The full script of this review can be found at my github page  [here](https://github.com/nunesmatt/tsbox-review/tsbox-review2.md).

---

The package is a comprehensive conversion tool for all current time series objects, stemming from the desire to be structure/class-agnostic.  As outlined in my first review, I feel this is a useful package in terms of functionality and scope for the R community.  The associated package vignettes and other documentation are well set out. Readdressing the comments from my first review:

Some comments on the package are below.  Note that these are suggestions rather than fixes per se:

1.  **Package checking**. When checking the package, I get:

```
dts_first_of_period: no visible binding for global variable ‘time.orig’
## Undefined global functions or variables:
##   time.orig
```

**This has been fixed successfully**.

2. **good practice**.  A few things came up when using `gp`.  Many code lines come up as 81 characters (as opposed to suggested lengths of $\leq 80$), which might be due to line endings etc.  `guess_dts.R` has a suggestion of using `vapply` rather than `sapply`.  These are left to the developer to tackle if desired.

@christophsax : You are right that these are SRR stats related, and as such I am fine with them being left as is.

3. **Error message about `ts_boxable(x)`**.  I wonder whether a more informative error message could be shown when `ts_boxable(x)` is not `TRUE`.

**This new error message is now more helpful to the beginner user**.

4.  The package developer might want to consider an additional function handling the `zooreg` class for completeness.

@christophsax : thanks for the inclusion.

5.  Whilst users are probably familiar with usage, the developer might want to consider adding examples to show functionality of `ts_prcomp` and `ts_forecast` under the `ts_examples` documentation file.

On reflection, including examples within `ts_plot` might indeed be overkill, and I suggest leaving the documentation as is.

6.  I noticed that there is a spelling mistake of *explicit* in the error messages for the `guess_time` and `guess_value` functions.

** this has been fixed successfully **.

Having run through various checks on the package, in summary @christophsax has addressed all my comments and has a great package for end-users.
