### 2023-01-20 #464 @mpadge
Thanks @nunesmatt. @christophsax it seems that these comments only leave a couple of very minor points:

1. >  `guess_dts.R` has a suggestion of using `vapply` rather than `sapply`.

    rOpenSci has a clear policy on using `vapply` and not `sapply`, so could you please change this?
2. Please comment on the error message for `ts_boxable()`, and optionally inform of any changes you may decide to implement.

Please let us know when these very minor and very final issues have been addressed, and we'll look forward to approving the package.

---

@chamberlinc Could you please indicate in the meantime your final approval of this package? Thank you, and thanks to all involved for your patience in this issue :smiley:
