### 2023-01-20 #464 @nunesmatt
@mpadge, bot of these issues have also been fixed by @christophsax in the latest update.  Apologies for my error in not including this information..
