### 2023-02-03 #464 @maurolepore
Dear all,

I don't mean to intrude but to save you time. My EiC rotation just started and I'm checking the status of open issues. So while I'm here I'll point to the next steps I see (handling editor feel free to correct me).

- Two reviewers submitted their reviews:
  -  [x] @nunesmatt expressed the approval explicitely.
  -  [ ] @chamberlinc could you please also express if you request more changes or approve ([approval template](https://devguide.ropensci.org/approval2template.html))

--

I don't see a review by @brunaw but I assume it's optional.
