### 2023-02-16 #464 @christophsax
@maurolepore, my understanding from https://github.com/ropensci/software-review/issues/464#issuecomment-1342939990 was that @chamberlinc has already approved.

Let me know if I can provide something else. Thanks a lot!
