### 2023-02-17 #464 @maurolepore
Thanks @christophsax.

I see no comment by @chamberlinc since you incorporated reviewers' feedback on Dec 26, 2022 (https://github.com/ropensci/software-review/issues/464#issuecomment-1365462292). But I see @mpadge sent a reminder for the final approval on Jan 20, 2023 (https://github.com/ropensci/software-review/issues/464#issuecomment-1398059879). With that and our mentions here I hope we get an answer soon :crossed_fingers:.
