### 2023-03-08 #464 @nunesmatt
> @chamberlinc and @nunesmatt, thank you very much for your helpful reviews! Is it ok for you if I add you as `["rev"]` to the `DESCRIPTION` in tsbox? Thanks!

@christophsax Sure! Thanks for the contribution to the package community.
